requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
aiohttp==3.9.5
discord.py==2.3.2
pydantic==2.6.0